requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.13.1
psycopg2-binary==2.9.10
bcrypt==4.3.0
PyJWT==2.10.1
